Add JSON output, refactor, prune bad edges, bugfix
Since the BlackMarket dump isn't 100% complete, and perhaps because of
deleted users, some edges refer to missing users. Those are now pruned.

JSON output is a simple serialization of users and edges.

Edge and user extraction had terrible copy-paste bug that resulted in
processing the dataset quadratically instead of linearly, oops!

Also, added a template.dot - it still doesn't help with regards to
visualizing the entire graph, but perhaps it's a starting point for
rendering a subset.